rd-456 improve search experience
