+++ packages/robindoc/src/components/ui/backdrop/backdrop.scss
@@ -4,24 +4,14 @@
   right: 0;
   bottom: 0;
   left: 0;
-  background: color-mix(in srgb, var(--r-main-900) 0%, transparent 100%);
   opacity: 0;
   visibility: hidden;
   pointer-events: none;
-  backdrop-filter: blur(4px);
-  transition:
-    background 0.05s,
-    opacity 0.2s,
-    visibility 0.01s 0.15s;
   z-index: 100;
 }
 
 .r-backdrop._visible {
   opacity: 1;
-  background: color-mix(in srgb, var(--r-main-900) 10%, transparent 100%);
   visibility: visible;
-  transition:
-    background 0.05s,
-    opacity 0.2s;
   pointer-events: all;
 }
