--- packages/robindoc/src/components/ui/modal/modal.scss
@@ -1,42 +1,0 @@
-@use "src/assets/_vars";
-
-.r-modal {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  background-color: color-mix(in srgb, var(--r-main-50) 95%, transparent);
-  border: 2px solid var(--r-main-200);
-  border-radius: 12px 12px 0 0;
-  opacity: 0;
-  visibility: hidden;
-  pointer-events: none;
-  transition:
-    opacity 0.2s 0.1s,
-    transform 0.2s 0.1s,
-    visibility 0.01s 0.2s 0.1s;
-  z-index: 1000;
-  box-sizing: border-box;
-
-  @media screen and (width >=vars.$sm) {
-    top: 10%;
-    left: 50%;
-    bottom: auto;
-    transform: translateX(-50%) scale(0.95);
-    max-width: 480px;
-    border-radius: 12px;
-  }
-}
-
-.r-modal._visible {
-  opacity: 1;
-  visibility: visible;
-  transition:
-    opacity 0.2s 0.1s,
-    transform 0.2s 0.1s;
-  pointer-events: all;
-
-  @media screen and (width >=vars.$sm) {
-    transform: translateX(-50%) scale(1);
-  }
-}
